cambios en rutas y mensajes de error

+++ src/app/vistas/start/start.component.ts
@@ -98,11 +98,10 @@ export class StartComponent implements OnInit {
   httpErrorType:number=0;
 /*    Errores:
     0: no hay error
-    1: error en la url o token invalido
-    2: timeout error
-    401: codigo de seguridad erroneo
+    500: timeout error
+    504: timeout error
     404: paciente no encontrado
-    500: paciente no encontrado
+    404: combinación paciente y accesion inexistente
 */
   loading:boolean=false;
   pacienteExiste:boolean=false;
@@ -126,17 +125,85 @@ export class StartComponent implements OnInit {
     this.api.getPacienteUrl(tokenUrl).subscribe(
       (data:PacienteInterface)=>{
         this.paciente=data;
-        if(this.paciente && (this.httpErrorType!=500)){
-          this.paciente=data;
-          this.pacienteExiste=true;
-        } else {
-          this.httpErrorMsg = this.api.httpErrorMsg;
-          this.httpErrorType = this.api.httpErrorType;
-          console.log(this.httpErrorMsg);
-          console.log(this.httpErrorType);
-        }
+        this.pacienteExiste=true;
+        this.httpErrorMsg = "";
+        this.httpErrorType = 0;
+        localStorage.setItem("beneficiario",this.paciente.apyNom);
+        localStorage.setItem("emailPaciente", this.paciente.emailActivo);
+        localStorage.setItem("pacienteId", this.paciente.pacienteId.toString());
+        localStorage.setItem("accession", this.accesoId.value);
         this.loading=false;
       }, error => {
+        this.paciente = {
+          pacienteId: 0,
+          nombres: "",
+          apellido: "",
+          apyNom: "",
+          fechaNacim: "",
+          sexo: "",
+          estadoCivilId: 0,
+          tipoDocumentoId: "",
+          tipoDocumentoIdDesc: "",
+          nroDocumento: "",
+          calle: "",
+          alturaCalle: "",
+          piso: "",
+          departamento: "",
+          localidadId: 0,
+          provinciaId: 0,
+          paisId: 0,
+          codPos: "",
+          ctaCte: "",
+          historiaClinica: 0,
+          ondicionIVAId: 0,
+          email: "",
+          emailActivo: "",
+          exentoIIBBPciaBsas: "",
+          tipoImpuestoIIBBId: 0,
+          impuestoIIBBPciaBsasId: 0,
+          sysInactivo: 0,
+          sysFechac: "",
+          sysLoginc: "",
+          sysHostc: "",
+          sysFecham: "",
+          sysLoginm: "",
+          sysHostm: "",
+          paisNacimientoId: 0,
+          beneficiario: "",
+          exentoIIBBCapfed: "",
+          impuestoIIBBCapfedId: 0,
+          mor: "",
+          telemedicina: "",
+          benefactor: "",
+          carpetaHc: "",
+          respApellido: "",
+          respNombre: "",
+          respTipoDocId: 0,
+          respNroDoc: "",
+          respEmail: "",
+          respSexo: "",
+          respParentescoId: 0,
+          respTel1: "",
+          respCalle: "",
+          respLocalidadId: 0,
+          respProvinciaId: 0,
+          respPaisId: 0,
+          respCodPos: "",
+          consentEstudioInternet: "",
+          alerta: "",
+          empadronado: "",
+          direccion: "",
+          habilitarPortal: "",
+          habilitarPortalResp: "",
+          telefonos: [
+              {
+                  telefonoId: "",
+                  numero: "",
+                  tipoTelefonoId: ""
+              }
+          ]
+        };
+        this.httpErrorType = this.api.httpErrorType;
         this.httpErrorMsg=this.api.httpErrorMsg;
         this.loading=false;
       });
@@ -144,34 +211,30 @@ export class StartComponent implements OnInit {
 
   sendCode(form: any){
     this.loading=true;
-    this.postPaciente.app='RIS';
-    this.postPaciente.pacienteId=form.pacienteId.toString();
-    this.postPaciente.id=form.accesoId.toString();
-    if (this.pacienteExiste){
-      console.log(this.postPaciente);
-      this.api.sendCode(this.postPaciente).subscribe(
-        (data:any) =>{
-        if(data){
-          localStorage.setItem("beneficiario",this.paciente.apyNom);
-          localStorage.setItem("emailPaciente", this.paciente.emailActivo);
-          localStorage.setItem("pacienteId", this.paciente.pacienteId.toString());
-          localStorage.setItem("accession", form.accesoId.toString());
-          localStorage.setItem("estudios",data.urlStudy);
-          this.router.navigate(['ris-link-send-mail']);
-        } else {
-          this.httpErrorMsg = data.msg;
-          this.httpErrorType=data.status;
-        }
-          this.loading=false;
-      }, error =>{
-          this.loading=false;
-          this.httpErrorMsg=this.api.httpErrorMsg;
-          this.httpErrorType=this.api.httpErrorType;
-      });
-    } else{
-      this.getPacienteUrl(form.pacienteId.toString());
-      this.pacienteExiste=true;
-      this.sendCode(form);
+    this.getPacienteUrl(form.pacienteId.toString());
+    if((this.pacienteExiste) && (this.httpErrorType==0)){
+      this.postPaciente.app='RIS';
+      this.postPaciente.pacienteId=form.pacienteId.toString();
+      this.postPaciente.id=form.accesoId.toString();
+        this.api.sendCode(this.postPaciente).subscribe(
+          (data:any) =>{
+            if(data){
+              localStorage.setItem("beneficiario",this.paciente.apyNom);
+              localStorage.setItem("emailPaciente", this.paciente.emailActivo);
+              localStorage.setItem("pacienteId", this.paciente.pacienteId.toString());
+              localStorage.setItem("accession", form.accesoId.toString());
+              localStorage.setItem("estudios",data.urlStudy);
+              this.router.navigate(['ris-link-send-mail']);
+            } else {
+              this.httpErrorMsg = data.msg;
+              this.httpErrorType=data.status;
+            }
+              this.loading=false;
+          }, error =>{
+              this.httpErrorMsg=this.api.httpErrorMsg;
+              this.httpErrorType=this.api.httpErrorType;
+              this.loading=false;
+          });
     }
   }
